✅(backend) add tests for joanie lms handler
Add a test to ensure that the course run synchronization is working with Joanie.
 Previously we only support Open EdX so all tests related to this api endpoint
 were grouped within a test case dedicated to Open EdX. As now, we also support
 Joanie, we split this test case into a generic test case, a test case for Open
 EdX and a last one for Joanie.
